+++ packages/components/src/shared/options.ts
@@ -1,7 +1,12 @@
 import type { LocaleConfig } from "@vuepress/core";
 import type { BackToTopLocaleData } from "./locales";
 
-export type AvailableComponent = "Badge" | "CodePen" | "FontIcon" | "PDF";
+export type AvailableComponent =
+  | "Badge"
+  | "CodePen"
+  | "FontIcon"
+  | "PDF"
+  | "StackBlitz";
 
 export interface ComponentOptions {
   /**
@@ -30,7 +35,7 @@ export interface ComponentOptions {
    *
    * 是否启用返回顶部按钮
    *
-   * @description 当设置为数字时，将会作为返回顶部按钮距离阈值（单位: 像素）
+   * @description 当设置为数字时，将会作为返回顶部按钮距离阈值 (单位: 像素)
    *
    * @default false
    */