+++ packages/theme/src/client/module/sidebar/composables/resolveConfig.ts
@@ -75,17 +75,6 @@ export const resolveArraySidebarItems = (
   const page = usePageData();
   const route = useRoute();
 
-  if (!Array.isArray(sidebarConfig)) {
-    console.warn(
-      `Expecting array, but getting invalid sidebar config${
-        prefix ? ` under ${prefix}` : ""
-      } with:`,
-      sidebarConfig
-    );
-
-    return [];
-  }
-
   const handleChildItem = (
     item: HopeThemeSidebarItem,
     pathPrefix = prefix
@@ -110,24 +99,10 @@ export const resolveArraySidebarItems = (
           ? sidebarData[prefix]
           : childItem.children;
 
-      if (!Array.isArray(children)) {
-        console.warn(
-          `Expecting array, but getting invalid sidebar config${
-            prefix ? ` under ${prefix}` : ""
-          } with:`,
-          children
-        );
-
-        return {
-          type: "group",
-          ...childItem,
-          children: [],
-        };
-      }
-
       return {
         type: "group",
         ...childItem,
+        prefix,
         children: children.map((item) => handleChildItem(item, prefix)),
       };
     }